requirements.txt
# Core Flask and Web Framework
blinker==1.9.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login>=0.6
Werkzeug==3.1.6
gunicorn==21.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
colorama==0.4.6

# HTTP and Networking
requests==2.32.5
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0

# Database
SQLAlchemy>=2.0
psycopg2-binary>=2.9

# Configuration
python-dotenv>=1.0

# Scientific Computing (Weather/Thermal Analysis)
numpy>=1.24

# QR Code Generation (XCTask format)
qrcode[pil]>=7.4
Pillow>=10.0
